% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{pipedrive_api_content}
\alias{pipedrive_api_content}
\title{pipedrive_api_content}
\usage{
pipedrive_api_content(.res, .as = "text")
}
\arguments{
\item{.res}{api response}

\item{.as}{parsing as text, json, ...}
}
\description{
pipedrive_api_content
}
