% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{pipedrive_api_deals}
\alias{pipedrive_api_deals}
\title{pipedrive_api_deals}
\usage{
pipedrive_api_deals(.url, .token, .start = 0, .limit = 500,
  .verbose = FALSE)
}
\arguments{
\item{.url}{api url}

\item{.token}{api token}

\item{.start}{pagination start}

\item{.limit}{pagination limit (500 max)}

\item{.verbose}{TRUE/FALSE}
}
\description{
pipedrive_api_deals
}
