% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{pipedrive_api_get}
\alias{pipedrive_api_get}
\title{pipedrive_api_get}
\usage{
pipedrive_api_get(.url, .path, .query, .verbose = FALSE)
}
\arguments{
\item{.url}{pipedrive url}

\item{.path}{api path}

\item{.query}{api query}

\item{.verbose}{TRUE/FALSE}
}
\description{
pipedrive_api_get
}
