% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{pipedrive_paginate}
\alias{pipedrive_paginate}
\title{pipedrive_paginate}
\usage{
pipedrive_paginate(.api, .start = 0, .limit = 500, .page_limit = 100)
}
\arguments{
\item{.api}{api functional}

\item{.start}{pagination start}

\item{.limit}{pagination limit (500 max)}

\item{.page_limit}{maximum number of downloaded pages}
}
\description{
pipedrive_paginate
}
